Extract the logger_patch and apply with prepend
This change also fixes the lack of names of the named parameters on
the fallback to the original Logger.initialize
