+++ test/unit/harness-suite.json
@@ -206,6 +206,23 @@
         }
     },
     {
+      "name": "test_logger_logs_to_a_file_when_logdev_is_a_file",
+      "handler": "log.log_to_file_and_read_back",
+      "environmentVariables": {
+        "_LAMBDA_TELEMETRY_LOG_FD": "test/unit/resources/fd/test_fd"
+      },
+      "request": {
+        "messages": ["Single frame written to a file\n even if there are multiple lines\nthird line"],
+        "fd_path": "test/unit/resources/fd/test_fd",
+        "file_path": "test/unit/resources/fd/test_file.txt"
+      },
+      "assertion":
+        {
+          "transform": "I, \\[[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])T(2[0-3]|[01][0-9]):[0-5][0-9]:[0-9]{2}.[0-9]{6} #[0-9]*\\]  INFO -- : Single frame written to a file\\\\n even if there are multiple lines\\\\nthird line",
+          "response": true
+        }
+    },
+    {
       "name": "test_logger_uses_telemetry_log_fd_when_logdev_is_stderr",
       "handler": "log.log_to_stderr_and_read_fd",
       "environmentVariables": {
